==> 2_Nettoyage_QR.ipynb <==
### Nettoyage des questions et des réponses :

import unicodedata 
import time
import pandas as pd
import numpy as np
import random
import nltk
import collections
import csv
import pickle
from pandas import DataFrame
from sklearn.feature_extraction.text import TfidfVectorizer # ou CountVectorizer
import re

data = pd.read_csv('dataQR.csv', sep=';')

data.iloc[1][1]

On remplace les retours à la ligne par des valeurs vides : 

nltk_stopwords = nltk.corpus.stopwords.words('french')
stopwords = list(nltk_stopwords)
stemmer=nltk.stem.SnowballStemmer('french')

def nettoyage_texte(txt):
    #on met tout le texte en minuscule, et on remplce les retours à la ligne par des espaces
    txt = txt.lower().replace('\n', ' ')

    #on retire tous les accents 
    txt = unicodedata.normalize('NFD', txt).encode('ascii', 'ignore')

    #on garde uniquement les caractères alphanumériques : 
    txt = re.sub('[^a-z_]', ' ', str(txt))
    
    #on retire les stops words
    tokens = [w for w in txt.split() if (len(w)>2) and (w not in nltk_stopwords)]
    
    #pour avoir les mots racines
    tokens = [stemmer.stem(token) for token in tokens]

    return ' '.join(tokens)

with open("nettoyage_texte.pk", "wb") as f:
    b = pickle.Pickler(f)
    b.dump(nettoyage_texte)

data['question'] = data['question'].apply(lambda x : nettoyage_texte(x))

data['thematiques'] = data['thematiques'].apply(lambda x : nettoyage_texte(x))

data['reponse'] = data['reponse'].apply(lambda x : nettoyage_texte(x))


data.iloc[1][1]

Création du Dataframe normalisé : 

dataQR_normalize = DataFrame({'Question_normalize' : data['question'], 'Reponse_normalize' : data['reponse'], 'Thematiques_normalize': data['thematiques']})

dataQR_normalize.to_csv('dataQR_normalize.csv', index = False, sep=';')

On construit un vectoriseur sur les questions :

vectorizer_q = TfidfVectorizer()

dtm_question = vectorizer_q.fit(data['question'])

with open('dtm_question.pk', 'wb') as fin:
    pickle.dump(vectorizer_q, fin)

#Affichage de la matrice : 
pd.DataFrame(dtm_question.transform(data['question']).toarray(), columns=vectorizer_q.get_feature_names())

On construit un vectoriseur sur les réponses : 

vectorizer_r = TfidfVectorizer()

dtm_reponse = vectorizer_r.fit(data['reponse'])

with open('dtm_reponse.pk', 'wb') as fin:
    pickle.dump(vectorizer_r, fin)

data['reponse']

#Affichage de la matrice : 
pd.DataFrame(dtm_reponse.transform(data['reponse']).toarray(), columns=vectorizer_r.get_feature_names())

On construit un vectoriseur sur les thématiques :

vectorizer_t = TfidfVectorizer()

dtm_thematiques = vectorizer_t.fit(data['thematiques'])

with open('dtm_thematiques.pk', 'wb') as fin:
    pickle.dump(vectorizer_t, fin)

#Affichage de la matrice : 
pd.DataFrame(dtm_thematiques.transform(data['thematiques']).toarray(), columns=vectorizer_t.get_feature_names())
